technical_server/pyproject.toml
[project]
name = "technical-server"
version = "0.1.0"
description = "Back end server for technical test"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pandas>=2.3.0",
    "psycopg>=3.2.6",
    "psycopg-pool>=3.2.6",
    "pydantic-settings>=2.8.1",
]

[dependency-groups]
dev = ["black>=25.1.0"]
